added statusCode & cachedAt to cache entries

+++ MONK/Internal/Classes/Cache.swift
@@ -8,32 +8,41 @@
 
 import Foundation
 
-class Cache {
+/// This controls the system's purge behavior
+///
+/// - purgeOnLowDiskSpace: iOS will be free to purge all/part of the cache on OS updates & low disk space situations. The cache is also purged when you manually call `emptyCache()` or `uncacheObject(for:)`, or when an object's expiration (if it has one) passes.
+/// - manualPurgeOrExpirationOnly: The cache is only purged when you manually call `emptyCache()` or `uncacheObject(for:)`, or when an object's expiration (if it has one) passes.
+public enum CacheBehavior {
+    case purgeOnLowDiskSpace
+    case manualPurgeOrExpirationOnly
+}
+
+public class Cache {
     fileprivate enum Constants {
         static let cacheFolderName = "MONK_Cache"
         static let cacheMetadataFileName = "cache.metadata"
     }
 
-    private let behavior: Behavior
+    private let behavior: CacheBehavior
     fileprivate let fileManager: FileManager
-    fileprivate let cacheFolderURL: URL
-    fileprivate let cacheMetadataFile: URL
+    fileprivate let cacheFolderURL: URL?
+    fileprivate let cacheMetadataFile: URL?
     fileprivate var cacheEntries: [URL : CacheEntry]
 
-    /// This controls the system's purge behavior
-    ///
-    /// - purgeOnLowDiskSpace: iOS will be free to purge all/part of the cache on OS updates & low disk space situations. The cache is also purged when you manually call `emptyCache()` or `uncacheObject(for:)`, or when an object's expiration (if it has one) passes.
-    /// - manualPurgeOrExpirationOnly: The cache is only purged when you manually call `emptyCache()` or `uncacheObject(for:)`, or when an object's expiration (if it has one) passes.
-    enum Behavior {
-        case purgeOnLowDiskSpace
-        case manualPurgeOrExpirationOnly
+    struct CacheData {
+        let data: Data
+        let cachedAt: Date
+        let statusCode: Int
+
     }
 
-    /// Create the cache. Generally you would want a maximum of 2 caches. One for each behavior, if you create duplicate caches for a behavior, they could interfeer with each other and cause issues.
+    static var purgableCache: Cache = Cache(behavior: .purgeOnLowDiskSpace)
+    static var persistantCache: Cache = Cache(behavior: .manualPurgeOrExpirationOnly)
+
+    /// Create the cache. Generally you would want a maximum of 2 caches. One for each behavior, if you create duplicate caches for a behavior, they could interfeer with each other and cause issues. Typically you would use `purgableCache` or `persistantCache` instead of this `init` to prevent that clashing.
     ///
     /// - Parameter behavior: Should the system be allowed to purge the cache or not
-    /// - Throws: FileManager errors if unable to create the base cache directory
-    init(behavior: Behavior) throws {
+    init(behavior: CacheBehavior) {
         self.fileManager = FileManager.default
         self.behavior = behavior
         let directory: FileManager.SearchPathDirectory
@@ -43,9 +52,9 @@ class Cache {
         case .manualPurgeOrExpirationOnly:
             directory = .applicationSupportDirectory
         }
-        let baseCacheURL = try fileManager.url(for: directory, in: .userDomainMask, appropriateFor: nil, create: true)
-        cacheFolderURL = baseCacheURL.appendingPathComponent(Constants.cacheFolderName, isDirectory: true)
-        cacheMetadataFile = cacheFolderURL.appendingPathComponent(Constants.cacheMetadataFileName)
+        let baseCacheURL = try? fileManager.url(for: directory, in: .userDomainMask, appropriateFor: nil, create: true)
+        cacheFolderURL = baseCacheURL?.appendingPathComponent(Constants.cacheFolderName, isDirectory: true)
+        cacheMetadataFile = cacheFolderURL?.appendingPathComponent(Constants.cacheMetadataFileName)
         cacheEntries = [:]
 
         createCacheDirectory()
@@ -53,7 +62,8 @@ class Cache {
     }
 
     /// Remove everything from the cache
-    func removeAll() {
+    public func removeAll() {
+        guard let cacheFolderURL = cacheFolderURL else { return }
         do {
             try fileManager.removeItem(at: cacheFolderURL)
         } catch {
@@ -87,8 +97,9 @@ class Cache {
     /// - Parameters:
     ///   - object: The object to cache
     ///   - url: The remote URL that was used to fetch the `object`
+    ///   - statusCode: The status code originally recieved when the object was fetched
     ///   - expiration: If you want this object to expire, this should be the time that expiration should occur. The cached object is not guarenteed to be purged as soon as this time is hit, but with normal cache use it should be purged shortly there after. You will never be able to get an expired object out of the cache. If the expiration is in the past, then the object won't be cached to begin with.
-    func add(object: Data, url: URL, expiration: Date?) {
+    func add(object: Data, url: URL, statusCode: Int, expiration: Date?) {
         // incase there is a previously cached entry for this URL, remove it first, otherwise it's cache file would become orphaned and would not be able to be individually purged.
         removeObject(for: url)
         if let expiration = expiration, expiration < Date() {
@@ -96,7 +107,7 @@ class Cache {
             return
         }
 
-        let entry = createCacheEntry(forRequestURL: url, expiration: expiration)
+        guard let entry = createCacheEntry(forRequestURL: url, statusCode: statusCode, expiration: expiration) else { return }
         do {
             try object.write(to: entry.cacheURL, options: .atomic)
             cacheEntries[url] = entry
@@ -110,13 +121,14 @@ class Cache {
     ///
     /// - Parameter url: The remote URL that you want the cached response data for
     /// - Returns: The cached response data, nil if there isn't any, or it had expired
-    func cachedObject(for url: URL) -> Data? {
+    func cachedObject(for url: URL) -> CacheData? {
         purgeAnyExpiredObjects()
         guard let entry = cacheEntries[url] else {
             return nil
         }
         do {
-            return try Data(contentsOf: entry.cacheURL)
+            let data = try Data(contentsOf: entry.cacheURL)
+            return CacheData(data: data, cachedAt: entry.cachedAt, statusCode: entry.statusCode)
         } catch {
             print("Couldn't read MONK cache file with error: \(error)")
             // We can't read the data from disk, so we should stop thinking we have this object cached
@@ -128,6 +140,7 @@ class Cache {
 
 private extension Cache {
     func createCacheDirectory() {
+        guard let cacheFolderURL = cacheFolderURL else { return }
         do {
             try fileManager.createDirectory(at: cacheFolderURL, withIntermediateDirectories: true, attributes: nil)
         } catch {
@@ -137,6 +150,7 @@ private extension Cache {
 
     func loadCacheEntries() {
         cacheEntries.removeAll()
+        guard let cacheMetadataFile = cacheMetadataFile else { return }
 
         do {
             let data = try Data(contentsOf: cacheMetadataFile)
@@ -152,6 +166,7 @@ private extension Cache {
     }
 
     func saveCacheEntries() {
+        guard let cacheMetadataFile = cacheMetadataFile else { return }
         let entries = cacheEntries.map({ $0.value.json })
         do {
             let data = try JSONSerialization.data(withJSONObject: entries, options: [])
@@ -181,11 +196,13 @@ private extension Cache {
         saveCacheEntries()
     }
 
-    func createCacheEntry(forRequestURL url: URL, expiration: Date?) -> CacheEntry {
+    func createCacheEntry(forRequestURL url: URL, statusCode: Int, expiration: Date?) -> CacheEntry? {
+        guard let cacheFolderURL = cacheFolderURL else { return nil }
+
         let uuid = UUID().uuidString
         let cacheFileName = "\(uuid).cache"
         let cacheURL = cacheFolderURL.appendingPathComponent(cacheFileName)
-        let entry = CacheEntry(cacheURL: cacheURL, requestURL: url, expiration: expiration)
+        let entry = CacheEntry(cacheURL: cacheURL, requestURL: url, statusCode: statusCode, expiration: expiration)
         return entry
     }
 }
